pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autogen-ecm-organoid-assistant"
version = "0.1.0"
description = "A personal AutoGen starter project for extracellular matrix and organoid synthesis research."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "autogen-agentchat>=0.7.4",
  "autogen-ext[openai]>=0.7.4",
  "httpx>=0.27.2",
  "matplotlib>=3.10.0",
  "pypdf>=5.4.0",
  "pywebview>=5.4",
  "python-dotenv>=1.0.1",
  "scipy>=1.16.0",
  "streamlit>=1.44.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
]
desktop-build = [
  "pyinstaller>=6.16.0",
]

[project.scripts]
ecm-organoid-agent = "ecm_organoid_agent.__main__:main"
ecm-organoid-ui = "ecm_organoid_agent.frontend:launch"
ecm-organoid-desktop = "ecm_organoid_agent.desktop:main"
ecm-organoid-demo = "ecm_organoid_agent.demo:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
